requirements.txt
beautifulsoup4
flask
requests
gunicorn
nba_api
python-dotenv